backend/requirements.txt
# backend/requirements.txt

Flask==2.2.5 # Or your desired Flask version
gunicorn==20.1.0 # For running Flask app in production on Linux/Docker
python-dotenv==1.0.1 # For loading .env files
cryptography==41.0.7 # For Fernet encryption
SQLAlchemy==2.0.25 # Database ORM
psycopg2-binary==2.9.9 # PostgreSQL database adapter for SQLAlchemy
minio==7.2.15 # MinIO client library (replaces boto3 for direct MinIO interaction)
kafka-python==1.4.7 # Kafka client library (replaces confluent-kafka for simplicity and broader compatibility)
Flask-Cors==4.0.0 # If you need CORS support for your API
celery==5.3.6 # Distributed task queue
redis==5.0.1 # Redis client for Celery broker/backend (Valkey is Redis-compatible)
six==1.16.0 # <--- ADD THIS LINE
Werkzeug==2.2.3 # Or a compatible version
Flask-JWT-Extended==4.4.0 # Puedes especificar una versión si lo deseas, o simplemente Flask-JWT-Extended
boto3
pypdf
requests
pgvector
